requirements.txt
streamlit==1.28.1
google-ai==0.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
pandas==2.1.3
markdown==3.5.1 